+++ Src/Library/changelog.md
@@ -1,2 +1,3 @@
 ### FIX
 - pre/post processor collection modification bug #224
+- response dto initialization not working with array types #225